Break circular refs
