+++ packages/core/src/index.ts
@@ -18,6 +18,7 @@ export * from "./name-policy.js";
 export * from "./refkey.js";
 export * from "./render.js";
 export * from "./stc.js";
+export * from "./sti.js";
 export * from "./tap.js";
 export * from "./utils.js";
 export * from "./write-output.js";
